requirements.txt
discord.py
requests
pymongo
dnspython
validators